+++ src/cydsp/_core_common.h
@@ -49,3 +49,7 @@ void bind_madronalib(nb::module_ &m);
 void bind_hisstools(nb::module_ &m);
 void bind_choc(nb::module_ &m);
 void bind_grainflow(nb::module_ &m);
+void bind_vafilters(nb::module_ &m);
+void bind_bloscillators(nb::module_ &m);
+void bind_fxdsp(nb::module_ &m);
+void bind_iirdesign(nb::module_ &m);